copy in files from next-auth-example
Had to fix some typescript errors
