+++ package.json
@@ -26,6 +26,7 @@
     "react-dom": "18.2.0"
   },
   "devDependencies": {
+    "@types/node": "18.0.4",
     "eslint": "8.19.0",
     "eslint-config-next": "12.2.2"
   }
