--- pages/api/auth/[...nextauth].js
@@ -1,28 +1,0 @@
-import NextAuth from "next-auth"
-import GithubProvider from "next-auth/providers/github"
-
-
-export default NextAuth({
-  // Configure one or more authentication providers
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID,
-      clientSecret: process.env.GITHUB_SECRET,
-    }),
-    // ...add more providers here
-  ],
-  callbacks: {
-  async jwt({ token, account }) {
-    // Persist the OAuth access_token to the token right after signin
-    if (account) {
-      token.accessToken = account.access_token
-    }
-    return token
-  },
-  async session({ session, token, user }) {
-    // Send properties to the client, like an access_token from a provider.
-    session.accessToken = token.accessToken
-    return session
-  }
-}
-})
