+++ pages/api/auth/[...nextauth].ts
@@ -1,0 +1,42 @@
+import NextAuth, { NextAuthOptions } from "next-auth"
+import GoogleProvider from "next-auth/providers/google"
+// import FacebookProvider from "next-auth/providers/facebook"
+// import GithubProvider from "next-auth/providers/github"
+// import TwitterProvider from "next-auth/providers/twitter"
+// import Auth0Provider from "next-auth/providers/auth0"
+// import AppleProvider from "next-auth/providers/apple"
+// import EmailProvider from "next-auth/providers/email"
+
+// For more information on each option (and a full list of options) go to
+// https://next-auth.js.org/configuration/options
+
+export const authOptions: NextAuthOptions = {
+  // Configure one or more authentication providers
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_ID,
+      clientSecret: process.env.GOOGLE_SECRET,
+    }),
+    // GithubProvider({
+    //   clientId: process.env.GITHUB_ID,
+    //   clientSecret: process.env.GITHUB_SECRET,
+    // }),
+    // ...add more providers here
+  ],
+  callbacks: {
+    async jwt({ token, account }) {
+      // Persist the OAuth access_token to the token right after signin
+      if (account) {
+        token.accessToken = account.access_token
+      }
+      return token
+    },
+    async session({ session, token, user }) {
+      // Send properties to the client, like an access_token from a provider.
+      session.accessToken = token.accessToken
+      return session
+    }
+  }
+}
+
+export default NextAuth(authOptions)
