+++ pages/api/restricted.js
@@ -1,7 +1,7 @@
 import { unstable_getServerSession } from "next-auth/next"
 import { authOptions } from "./api/auth/[...nextauth]"
 
-export default async (req, res) => {
+export default async function restricted(req, res) {
   const session = await unstable_getServerSession(req, res, authOptions)
   if (session) {
     res.send({
